+++ packages/react/src/components/DataTable/TableToolbarMenu.tsx
@@ -5,14 +5,34 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import { Settings } from '@carbon/icons-react';
 import cx from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
-import OverflowMenu from '../OverflowMenu';
-import { Settings } from '@carbon/icons-react';
 import { usePrefix } from '../../internal/usePrefix';
+import OverflowMenu from '../OverflowMenu';
 
-const TableToolbarMenu = ({
+export interface TableToolbarMenuProps
+  extends React.HTMLAttributes<HTMLDivElement> {
+  children: React.ReactNode;
+
+  /**
+   * Provide an optional class name for the toolbar menu
+   */
+  className?: string;
+
+  /**
+   * The description of the menu icon.
+   */
+  iconDescription: string;
+
+  /**
+   * Optional prop to allow overriding the default menu icon
+   */
+  renderIcon?: React.ReactNode;
+}
+
+const TableToolbarMenu: React.FC<TableToolbarMenuProps> = ({
   className,
   renderIcon,
   iconDescription,
